% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GILKStest.R
\name{GILKStest}
\alias{GILKStest}
\title{Performing Kolmogorov-Smirnov Goodness of Fit Test for Distribution Generalized Inverse-Lindley}
\usage{
GILKStest(x, gamaawal, alphacv = 0.05, eps = 1e-06)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Performing Kolmogorov-Smirnov Goodness of Fit Test for Distribution Generalized Inverse-Lindley
}
